dsadsa

+++ src/api/courses.js
@@ -12,13 +12,13 @@ const router = Router();
 /*
  * Route to create a new course.
  */
-router.post('/', authenticateToken, authorizeRole('admin'), async function (req, res, next) {
+router.post('/courses', authenticateToken, authorizeRole('admin'), async function (req, res, next) {
   try {
     const course = await Course.create(req.body, CourseClientFields);
     res.status(201).send({ id: course.id });
   } catch (e) {
     if (e instanceof ValidationError) {
-      res.status(400).send({ error: e.message });
+      res.status(400).json({ error: 'The request body was either not present or did not contain a valid Course object.' });
     } else {
       next(e);
     }
@@ -28,7 +28,7 @@ router.post('/', authenticateToken, authorizeRole('admin'), async function (req,
 /*
  * Route to fetch the list of all courses (paginated).
  */
-router.get('/', async function (req, res, next) {
+router.get('/courses', async function (req, res, next) {
   const { page = 1, pageSize = 10, subject, number, term } = req.query;
   const whereClause = {};
 
@@ -62,14 +62,14 @@ router.get('/', async function (req, res, next) {
 /*
  * Route to fetch info about a specific course.
  */
-router.get('/:courseId', async function (req, res, next) {
+router.get('/courses/:courseId', async function (req, res, next) {
   const courseId = req.params.courseId;
   try {
     const course = await Course.findByPk(courseId);
     if (course) {
       res.status(200).send(course);
     } else {
-      res.status(404).send({ error: 'Course not found' });
+      res.status(404).json({ error: 'Specified Course `id` not found.' });
     }
   } catch (e) {
     next(e);
@@ -79,15 +79,15 @@ router.get('/:courseId', async function (req, res, next) {
 /*
  * Route to update a course.
  */
-router.patch('/:courseId', authenticateToken, async function (req, res, next) {
+router.patch('/courses/:courseId', authenticateToken, async function (req, res, next) {
   const courseId = req.params.courseId;
   try {
     const course = await Course.findByPk(courseId);
     if (!course) {
-      return res.status(404).send({ error: 'Course not found' });
+      return res.status(404).json({ error: 'Specified Course `id` not found.' });
     }
     if (req.user.role !== 'admin' && req.user.id !== course.instructorId) {
-      return res.status(403).send({ error: 'Unauthorized' });
+      return res.status(403).json({ error: 'The request was not made by an authenticated User satisfying the authorization criteria described above.' });
     }
     const result = await Course.update(req.body, {
       where: { id: courseId },
@@ -96,7 +96,7 @@ router.patch('/:courseId', authenticateToken, async function (req, res, next) {
     if (result[0] > 0) {
       res.status(200).send(course);
     } else {
-      res.status(404).send({ error: 'Course not found' });
+      res.status(400).json({ error: 'The request body was either not present or did not contain any fields related to Course objects.' });
     }
   } catch (e) {
     next(e);
@@ -106,14 +106,14 @@ router.patch('/:courseId', authenticateToken, async function (req, res, next) {
 /*
  * Route to delete a course.
  */
-router.delete('/:courseId', authenticateToken, authorizeRole('admin'), async function (req, res, next) {
+router.delete('/courses/:courseId', authenticateToken, authorizeRole('admin'), async function (req, res, next) {
   const courseId = req.params.courseId;
   try {
     const result = await Course.destroy({ where: { id: courseId }});
     if (result > 0) {
       res.status(204).send();
     } else {
-      res.status(404).send({ error: 'Course not found' });
+      res.status(404).json({ error: 'Specified Course `id` not found.' });
     }
   } catch (e) {
     next(e);
@@ -123,7 +123,7 @@ router.delete('/:courseId', authenticateToken, authorizeRole('admin'), async fun
 /*
  * Route to fetch students enrolled in a course.
  */
-router.get('/:courseId/students', authenticateToken, async function (req, res, next) {
+router.get('/courses/:courseId/students', authenticateToken, async function (req, res, next) {
   const courseId = req.params.courseId;
   try {
     const course = await Course.findByPk(courseId, {
@@ -135,11 +135,11 @@ router.get('/:courseId/students', authenticateToken, async function (req, res, n
     });
     if (course) {
       if (req.user.role !== 'admin' && req.user.id !== course.instructorId) {
-        return res.status(403).send({ error: 'Unauthorized' });
+        return res.status(403).json({ error: 'The request was not made by an authenticated User satisfying the authorization criteria described above.' });
       }
       res.status(200).send(course.students);
     } else {
-      res.status(404).send({ error: 'Course not found' });
+      res.status(404).json({ error: 'Specified Course `id` not found.' });
     }
   } catch (e) {
     next(e);
@@ -149,20 +149,29 @@ router.get('/:courseId/students', authenticateToken, async function (req, res, n
 /*
  * Route to update enrollment for a course.
  */
-router.post('/:courseId/students', authenticateToken, async function (req, res, next) {
+router.post('/courses/:courseId/students', authenticateToken, async function (req, res, next) {
   const courseId = req.params.courseId;
-  const { studentIds } = req.body;
+  const { add, remove } = req.body;
   try {
     const course = await Course.findByPk(courseId);
     if (!course) {
-      return res.status(404).send({ error: 'Course not found' });
+      return res.status(404).json({ error: 'Specified Course `id` not found.' });
     }
     if (req.user.role !== 'admin' && req.user.id !== course.instructorId) {
-      return res.status(403).send({ error: 'Unauthorized' });
+      return res.status(403).json({ error: 'The request was not made by an authenticated User satisfying the authorization criteria described above.' });
     }
-    await CourseEnrollments.destroy({ where: { courseId } });
-    const enrollments = studentIds.map((studentId) => ({ courseId, userId: studentId }));
-    await CourseEnrollments.bulkCreate(enrollments);
+    if (!add && !remove) {
+      return res.status(400).json({ error: 'The request body was either not present or did not contain the fields described above.' });
+    }
+
+    if (add && add.length > 0) {
+      const enrollments = add.map((studentId) => ({ courseId, userId: studentId }));
+      await CourseEnrollments.bulkCreate(enrollments);
+    }
+    if (remove && remove.length > 0) {
+      await CourseEnrollments.destroy({ where: { courseId, userId: remove } });
+    }
+
     res.status(200).send({ message: 'Enrollments updated' });
   } catch (e) {
     next(e);
@@ -172,7 +181,7 @@ router.post('/:courseId/students', authenticateToken, async function (req, res, 
 /*
  * Route to fetch the roster of a course in CSV format.
  */
-router.get('/:courseId/roster', authenticateToken, async function (req, res, next) {
+router.get('/courses/:courseId/roster', authenticateToken, async function (req, res, next) {
   const courseId = req.params.courseId;
   try {
     const course = await Course.findByPk(courseId, {
@@ -184,7 +193,7 @@ router.get('/:courseId/roster', authenticateToken, async function (req, res, nex
     });
     if (course) {
       if (req.user.role !== 'admin' && req.user.id !== course.instructorId) {
-        return res.status(403).send({ error: 'Unauthorized' });
+        return res.status(403).json({ error: 'The request was not made by an authenticated User satisfying the authorization criteria described above.' });
       }
       const parser = new Parser({ fields: ['id', 'username'] });
       const csv = parser.parse(course.students);
@@ -192,7 +201,7 @@ router.get('/:courseId/roster', authenticateToken, async function (req, res, nex
       res.attachment(`roster_course_${courseId}.csv`);
       res.status(200).send(csv);
     } else {
-      res.status(404).send({ error: 'Course not found' });
+      res.status(404).json({ error: 'Specified Course `id` not found.' });
     }
   } catch (e) {
     next(e);
@@ -202,7 +211,7 @@ router.get('/:courseId/roster', authenticateToken, async function (req, res, nex
 /*
  * Route to fetch assignments for a course.
  */
-router.get('/:courseId/assignments', async function (req, res, next) {
+router.get('/courses/:courseId/assignments', async function (req, res, next) {
   const courseId = req.params.courseId;
   try {
     const course = await Course.findByPk(courseId, {
@@ -214,7 +223,7 @@ router.get('/:courseId/assignments', async function (req, res, next) {
     if (course) {
       res.status(200).send(course.assignments);
     } else {
-      res.status(404).send({ error: 'Course not found' });
+      res.status(404).json({ error: 'Specified Course `id` not found.' });
     }
   } catch (e) {
     next(e);
